requirements.txt
google-ads>=23.0.0
gspread>=6.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.0.0
